requirements.txt
twikit>=2.0.0
openai>=1.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
langdetect>=1.0.9
rich>=13.0.0
